Bots/HObot/requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
discord==1.0.1
discord.py==1.3.2
docopt==0.6.2
feedparser==5.2.1
Flask==1.1.2
gast==0.3.3
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googlefinance==0.7
grpcio==1.27.2
h5py==2.10.0
idna==2.9
IMDbPY==6.8
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
multidict==4.7.5
multitasking==0.0.9
nltk==3.4.5
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.3
pandas-datareader==0.8.1
pipreqs==0.4.10
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.15
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post2
tensorflow==2.2.0rc2
tensorflow-estimator==2.2.0rc0
termcolor==1.1.0
textblob==0.15.3
textgenrnn==2.0.0
tqdm==4.43.0
urllib3==1.25.8
websockets==8.1
Werkzeug==1.0.0
wrapt==1.12.1
yarg==0.1.9
yarl==1.4.2
yfinance==0.1.54
youtube-dl==2020.3.24
